+++ projects/ontimize-web-ngx/src/lib/components/bar-menu/o-bar-menu.component.ts
@@ -1,5 +1,6 @@
 import { Component, ElementRef, Injector, OnInit, ViewEncapsulation, forwardRef } from '@angular/core';
 
+import { Subscription } from 'rxjs';
 import { AppMenuService } from '../../services/app-menu.service';
 import { PermissionsService } from '../../services/permissions/permissions.service';
 import { OTranslateService } from '../../services/translate/o-translate.service';
@@ -35,6 +36,7 @@ export class OBarMenuComponent implements OnInit {
   protected _menuTitle: string;
   protected _tooltip: string;
   protected _id: string;
+  protected subscription: Subscription = new Subscription();
 
   constructor(
     protected elRef: ElementRef,
@@ -43,6 +45,7 @@ export class OBarMenuComponent implements OnInit {
     this.permissionsService = this.injector.get(PermissionsService);
     this.translateService = this.injector.get(OTranslateService);
     this.appMenuService = this.injector.get(AppMenuService);
+    this.subscription.add(this.appMenuService.onPermissionMenuChanged.subscribe(() => this.refreshMenuRoots()));
     this.menuRoots = this.appMenuService.getMenuRoots();
   }
 
@@ -110,4 +113,14 @@ export class OBarMenuComponent implements OnInit {
   get menuItems(): MenuRootItem[] {
     return this.menuRoots;
   }
+
+  refreshMenuRoots(): void {
+    this.menuRoots = [...this.appMenuService.getMenuRoots()];
+  }
+
+  ngOnDestroy(): void {
+    this.subscription.unsubscribe();
+
+  }
 }
+