+++ projects/ontimize-web-ngx/src/lib/types/table/o-table-global-config.type.ts
@@ -1,4 +1,10 @@
+import { ORowHeight, OTableDetailMode, OTableEditionMode } from './../../util/codes';
+
 export type OTableGlobalConfig = {
   autoAdjust: boolean;
   autoAlignTitles: boolean;
-};
+  filterColumnActiveByDefault: boolean;
+  editionMode: OTableEditionMode;
+  detailMode: OTableDetailMode;
+  rowHeight: ORowHeight
+}